pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cliforapi"
version = "0.1.0"
description = "Universal CLI that reads any OpenAPI spec and exposes endpoints as shell commands — optimized for AI agent consumption"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
cliforapi = "cliforapi.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cliforapi"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "respx>=0.21",
    "pytest-asyncio>=0.23",
]
